LayoutTests/imported/w3c/web-platform-tests/referrer-policy/same-origin/http-rp/same-origin/http-http/shared-worker/keep-origin-redirect/same-origin-insecure.http-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: SanityChecker

Harness Error (FAIL), message = ReferenceError: Can't find variable: SanityChecker
